+++ Language/Python/palindrome.py
